% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.r
\name{flowchart}
\alias{flowchart}
\title{flowchart}
\usage{
flowchart(dat, gp = gpar(fill = "lightgrey"), group_exc_shift = 0.2,
  arrow_obj = getOption("connectGrobArrow", default = arrow(ends =
  "last", type = "closed", length = unit(0.25, "cm"))),
  term_arrow_type = "L")
}
\arguments{
\item{dat}{dataframe with variables \code{level}, \code{group}, 
\code{incexc}, \code{text} and, optionally, \code{x}, \code{y}, 
\code{width}, \code{just}}

\item{gp}{colour to fill boxes with (single value)}

\item{group_exc_shift}{how much to move the exclusion group by}

\item{arrow_obj}{getOption object to define the arrow type}

\item{term_arrow_type}{terminal arrow type (only used for terminal 
exclusions). Can take one of \code{v}, \code{h}, \code{L}, \code{-}, 
\code{Z} or \code{N}}
}
\value{
dataframe containing \code{level}, \code{group}, \code{text}, 
\code{incexc}, \code{x}, \code{y}, \code{width}, which can be used to 
optimize the flowchart
}
\description{
flowchart
}
\examples{
dat <- data.frame(level = c(1, 2, 3, 4, 4, 5, 5, 6, 6),
group = c(NA, NA, NA, 1, 2, 1, 2, 1, 2),
text  = c("Assessed for eligibility", "Not eligible\\nfor inclusion", 
"Randomized", "Group1", "Group2", "Excl Group1", "Exc\\nGroup2", "Analysis1", 
"Analysis2"),
incexc = c("inc", "exc", "inc", "inc", "inc", "exc", "exc", "inc", "inc")
)
View(dat)
flowchart(dat)
}
