% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_percentage}
\alias{format_percentage}
\title{format_percentage}
\usage{
format_percentage(value, round_digits, trim_trailing_zeros)
}
\arguments{
\item{value}{numeric; percentage value to be rounded}

\item{round_digits}{integer; number of digits past the decimal to round to}

\item{trim_trailing_zeros}{logical; trim trailing zeros or not in returned value.}
}
\description{
rounds percentage values while allowing the option to `trim_trailing_zeros`
}
\keyword{internal}
