% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_prettyNum}
\alias{info_prettyNum}
\title{info_prettyNum}
\usage{
info_prettyNum(x, big.mark)
}
\arguments{
\item{x}{an atomic numerical or character object, possibly complex, typically a vector of real numbers.}

\item{big.mark}{character. Used to specify the thousands separator for patient count values.}
}
\description{
Wrapper for `prettyNum()` that provides a more informative warning message if user `big.mark` equal to the character defined in the R environment `OutDec` option.
}
\keyword{internal}
