% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_class}
\alias{is_class}
\title{is_class}
\usage{
is_class(x, class)
}
\arguments{
\item{x}{element to check}

\item{class}{desired class to check}
}
\description{
Function to check if an object is from a given class.
}
