% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fc}
\alias{plot.fc}
\title{plot.fc}
\usage{
\method{plot}{fc}(x, ...)
}
\arguments{
\item{x}{A \code{fc} object created using flowchart functions}

\item{...}{Arguments passed to \link{fc_draw}.}
}
\description{
plot method for \code{fc} object. It is a wrapper for the \link{fc_draw} function.
}
