% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fc}
\alias{print.fc}
\title{print.fc}
\usage{
\method{print}{fc}(x, ...)
}
\arguments{
\item{x}{A \code{fc} object created using flowchart functions}

\item{...}{Not used}
}
\description{
print method for \code{fc} object
}
