% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet_prettyNum}
\alias{quiet_prettyNum}
\title{quiet_prettyNum}
\usage{
quiet_prettyNum(x, big.mark)
}
\arguments{
\item{x}{an atomic numerical or character object, possibly complex, typically a vector of real numbers.}

\item{big.mark}{character. Used to specify the thousands separator for patient count values.}
}
\description{
Wrapper for `prettyNum()` that suppresses the original warning message if user `big.mark` equal to the character defined in the R environment `OutDec` option.
}
\keyword{internal}
