% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fc}
\alias{summary.fc}
\title{summary.fc}
\usage{
\method{summary}{fc}(object, ...)
}
\arguments{
\item{object}{À \code{fc} object created using flowchart functions}

\item{...}{Not used}
}
\description{
summary method for \code{fc} object
}
