% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_numbers}
\alias{update_numbers}
\title{update_numbers}
\usage{
update_numbers(object, big.mark = "")
}
\arguments{
\item{object}{fc object that we want to draw.}

\item{big.mark}{character. Used to specify the thousands separator for patient count values. Defaults is no separator (`""`); if not empty used as mark between every 3 digits (ex: `big.mark = ","` results in `1,000` instead of `1000`).}
}
\description{
Updates values of `n` or `N` referenced in the `text` column when user specifies `big.mark` argument in `fc_draw`.
}
\keyword{internal}
