% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_x}
\alias{update_x}
\title{update_x}
\usage{
update_x(x, i, n)
}
\arguments{
\item{x}{old horizontal position of the boxes}

\item{i}{position of the flowchart within the list of flowcharts.}

\item{n}{number of total flowcharts to merge}
}
\description{
Function to update the horizontal position of the previuosly created flow charts, to merge two flowcharts used in fc_merge().
}
\keyword{internal}
