% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_y}
\alias{update_y}
\title{update_y}
\usage{
update_y(y, type, x)
}
\arguments{
\item{y}{old height of the boxes}

\item{type}{type of box}

\item{x}{horizontal position of the boxes}
}
\description{
Function to update the vertical position of the previuosly created boxes, to add the new ones with fc_split() or fc_filter().
}
\keyword{internal}
