% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_y_stack}
\alias{update_y_stack}
\title{update_y_stack}
\usage{
update_y_stack(y, x, type, i, n)
}
\arguments{
\item{y}{old vertical position of the boxes}

\item{x}{old horizontal position of the boxes}

\item{type}{type of the boxes}

\item{i}{position of the flowchart within the list of flowcharts.}

\item{n}{number of total flowcharts to merge}
}
\description{
Function to update the vertical position of the previuosly created flow charts, to stack the two flowcharts with fc_stack().
}
