% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_y_stack_unite}
\alias{update_y_stack_unite}
\title{update_y_stack_unite}
\usage{
update_y_stack_unite(y, x, type)
}
\arguments{
\item{y}{old vertical position of the boxes}

\item{x}{old horizontal position of the boxes}

\item{type}{type of the boxes}
}
\description{
Function to update the vertical position of the previuosly created flowcharts, to stack the two flowcharts with fc_stack(), when `unite` is TRUE.
}
\keyword{internal}
