% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_filter.R
\name{fc_filter}
\alias{fc_filter}
\title{fc_filter}
\usage{
fc_filter(
  object,
  filter = NULL,
  N = NULL,
  label = NULL,
  text_pattern = "{label}\\n {n} ({perc}\%)",
  perc_total = FALSE,
  show_exc = FALSE,
  direction_exc = "right",
  label_exc = "Excluded",
  text_pattern_exc = "{label}\\n {n} ({perc}\%)",
  sel_group = NULL,
  round_digits = 2,
  just = "center",
  text_color = "black",
  text_fs = 8,
  text_fface = 1,
  text_ffamily = NA,
  text_padding = 1,
  bg_fill = "white",
  border_color = "black",
  just_exc = "center",
  text_color_exc = "black",
  text_fs_exc = 6,
  text_fface_exc = 1,
  text_ffamily_exc = NA,
  text_padding_exc = 1,
  bg_fill_exc = "white",
  border_color_exc = "black",
  offset_exc = NULL
)
}
\arguments{
\item{object}{fc object that we want to filter.}

\item{filter}{Expression that returns a logical value and are defined in terms of the variables in the data frame. The data base will be filtered by this expression, and it will create a box showing the number of rows satisfying this condition.}

\item{N}{Number of rows after the filter in case `filter` is NULL.}

\item{label}{Character that will be the title of the box. By default it will be the evaluated condition.}

\item{text_pattern}{Structure that will have the text in each of the boxes. It recognizes label, n, N and perc within brackets. For default it is "\{label\}\\n \{n\} (\{perc\}\%)".}

\item{perc_total}{logical. Should percentages be calculated using the total number of rows at the beginning of the flowchart? Default is FALSE, meaning that they will be calculated using the number at the parent leaf.}

\item{show_exc}{Logical value. If TRUE a box showing the number of excluded rows will be added to the flow chart.}

\item{direction_exc}{One of "left" or "right" indicating if the exclusion box goes into the left direction or in the right direction. By default is "right".}

\item{label_exc}{Character that will be the title of the added box showing the excluded patients. By default it will show "Excluded".}

\item{text_pattern_exc}{Structure that will have the text in each of the excluded boxes. It recognizes label, n, N and perc within brackets. For default it is "\{label\}\\n \{n\} (\{perc\}\%)".}

\item{sel_group}{Specify if the filtering has to be done only by one of the previous groups. By default is NULL.}

\item{round_digits}{Number of digits to round percentages. It is 2 by default.}

\item{just}{Justification for the text: left, center or right. Default is center.}

\item{text_color}{Color of the text. It is black by default. See the `col` parameter for \code{\link{gpar}}.}

\item{text_fs}{Font size of the text. It is 8 by default. See the `fontsize` parameter for \code{\link{gpar}}.}

\item{text_fface}{Font face of the text. It is 1 by default. See the `fontface` parameter for \code{\link{gpar}}.}

\item{text_ffamily}{Changes the font family of the text. Default is NA. See the `fontfamily` parameter for \code{\link{gpar}}.}

\item{text_padding}{Changes the text padding inside the box. Default is 1. This number has to be greater than 0.}

\item{bg_fill}{Box background color. It is white by default. See the `fill` parameter for \code{\link{gpar}}.}

\item{border_color}{Box border color. It is black by default. See the `col` parameter for \code{\link{gpar}}.}

\item{just_exc}{Justification for the text of the exclude box: left, center or right. Default is center.}

\item{text_color_exc}{Color of the text of the exclude box. It is black by default. See `text_color`.}

\item{text_fs_exc}{Font size of the text of the exclude box. It is 6 by default. See `text_fs`.}

\item{text_fface_exc}{Font face of the text of the exclude box. It is 1 by default. See the `fontface` parameter for \code{\link{gpar}}. See `text_fface`.}

\item{text_ffamily_exc}{Changes the font family of the text of the exclude box. Default is NA. See the `fontfamily` parameter for \code{\link{gpar}}. See `text_ffamily`.}

\item{text_padding_exc}{Changes the text padding inside the exclude box. Default is 1. This number has to be greater than 0.}

\item{bg_fill_exc}{Exclude box background color. It is white by default. See `bg_fill`.}

\item{border_color_exc}{Box background color of the exclude box. It is black by default. See `border_color`.}

\item{offset_exc}{Amount of space to add to the distance between the box and the excluded box (in the x coordinate). If positive, this distance will be larger. If negative, it will be smaller. This number has to be at least between 0 and 1 (plot limits) and the resulting x coordinate cannot exceed these plot limits. The default is NULL (no offset).}
}
\value{
List with the filtered dataset and the flowchart parameters with the resulting filtered box.
}
\description{
This function allows to filter the flowchart in function of a expression that returns a logic value that are defined in terms of the variables in the database. It will generate one box per group showing the number of rows of the group that matches the condition, and will retain only those rows in the data base.
}
\examples{
safo |>
  as_fc(label = "Patients assessed for eligibility") |>
  fc_filter(!is.na(group), label = "Randomized", show_exc = TRUE) |>
  fc_draw()

}
