% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_split.R
\name{fc_split}
\alias{fc_split}
\title{fc_split}
\usage{
fc_split(
  object,
  var,
  label = NULL,
  text_pattern = "{label}\\n {n} ({perc}\%)",
  sel_group = NULL,
  na.rm = FALSE,
  just = "center",
  text_color = "black",
  text_fs = 8,
  bg_fill = "white",
  border_color = "black"
)
}
\arguments{
\item{object}{fc object that we want to split.}

\item{var}{variable column of the database from which it will be splitted.}

\item{label}{Vector of characters with the label of each category in order. It has to have as many elements as categories has the column. By default, it will put the labels of the categories.}

\item{text_pattern}{Structure that will have the text in each of the boxes. It recognizes label, n, N and perc within brackets. For default it is "\{label\}\\n \{n\} (\{perc\}\%)".}

\item{sel_group}{Specify if the splitting has to be done only by one of the previous groups. By default is NULL.}

\item{na.rm}{logical. Should missing values of the grouping variable be removed? Default is FALSE.}

\item{just}{Justification for the text: left, center or right. Default is center.}

\item{text_color}{Color of the text. It is black by default.}

\item{text_fs}{Font size of the text. It is 8 by default.}

\item{bg_fill}{Box background color. It is white by default.}

\item{border_color}{Box border color. It is black by default.}
}
\value{
List with the dataset grouped by the splitting variable and the flowchart parameters with the resulting split.
}
\description{
This function allows to split the flowchart in function of the categories of a column of the database. It will generate as many boxes as categories has the column showing in each one the frequency of each category. It will additionally group the database per this column.
}
\examples{
clinic_patient |>
  dplyr::filter(!is.na(group)) |>
  as_fc(label = "Patients included") |>
  fc_split(group) |>
  fc_draw()

}
