% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_num_in_expr}
\alias{replace_num_in_expr}
\title{replace_num_in_expr}
\usage{
replace_num_in_expr(expr, row, big.mark)
}
\arguments{
\item{expr}{expression in `fc$text`.}

\item{row}{A row from the `fc` object containing `n`, `N`, and `perc` values.}

\item{big.mark}{character. Used to specify the thousands separator for patient count values.}
}
\description{
Helper function for `update_numbers()`.
}
\keyword{internal}
